config: eslint 공통 정리
